fix bug in sens_exact
we were sampling incorrectly from the minima.  we were trying to draw a
vector uniformly from an k-dimension hypersphere, but we were doing it
wrong.  It is done correctly in nested_sampling.models.harmonic_nowalk,
I simply coppied what was done there.

Also, start to add a test that uses nested sampling to compute the cv
of the HSA.